modify chainId type

+++ pallets/chainbridge/src/lib.rs
@@ -25,7 +25,7 @@ mod tests;
 const DEFAULT_RELAYER_THRESHOLD: u32 = 1;
 const MODULE_ID: PalletId = PalletId(*b"cb/bridg");
 
-pub type ChainId = u8;
+pub type ChainId = u32;
 pub type DepositNonce = u64;
 pub type ResourceId = [u8; 32];
